refactor(RequestDto): 요청 DTO들의 package를 이동시켰습니다
Request관련 DTO -> application 패키지로 이동
- request 관련 DTO는 service 로직에서 사용하기 때문에, application 패키지로 이동하였습니다.
- application 패키지에 service로직을 둔 이유는 application 계층이 비즈니스 계층과 연결해주는 역할을 한다고 알고 있기 때문입니다. (제가 틀렸을 시 마구 지적해주세요)
- 제가 생각하는 Service 로직은 도메인과 컨트롤러(presentation) 계층을 이어주는 역할을 하고있기 때문에 application layer에 두었습니다.
