+++ src/main/java/com/realworld/study/post/presentation/dto/PostDeleteResponse.java
@@ -1,0 +1,4 @@
+package com.realworld.study.post.presentation.dto;
+
+public record PostDeleteResponse(boolean isDeleted) {
+}
